+++ crates/zng-app/src/widget/node/arc.rs
@@ -175,7 +175,7 @@ impl WeakNode {
     }
 }
 
-use parking_lot::Mutex;
+use zng_task::parking_lot::Mutex;
 
 use super::UiNode;